+++ changelog.md
@@ -7,6 +7,8 @@
 - Add debouncing for fuzzy file preview using `SingleAlarm`
 - Refactor everything
 - Add auto sizing option for the popup (default)
+- Add new actions FuzzyGrepOpenTabsCI, FuzzyGrepOpenTabs, FuzzyGrepCurrentBufferCI, FuzzyGrepCurrentBuffer
+- Rename an existing action from FuzzyGrepCaseInsensitive to FuzzyGrepCI (the old version will still work for now)
 
 ## Version 1.14.0